break out observer and fix bugs
